move compact_block and compact_transaction into p2p crate
