+++ core/src/core/mod.rs
@@ -16,8 +16,6 @@
 
 pub mod block;
 pub mod committed;
-pub mod compact_block;
-pub mod compact_transaction;
 pub mod hash;
 pub mod id;
 pub mod merkle_proof;
@@ -35,8 +33,6 @@ use util::secp::pedersen::Commitment;
 
 pub use self::block::*;
 pub use self::committed::Committed;
-pub use self::compact_block::*;
-pub use self::compact_transaction::*;
 pub use self::id::ShortId;
 pub use self::transaction::*;
 use core::hash::Hashed;
