+++ core/tests/block.rs
@@ -23,15 +23,13 @@ use std::time::Instant;
 
 pub mod common;
 
-use chrono::Duration;
 use common::{new_block, tx1i2o, tx2i1o, txspend1i1o};
 use grin_core::consensus::{self, BLOCK_OUTPUT_WEIGHT, MAX_BLOCK_WEIGHT};
 use grin_core::core::block::Error;
 use grin_core::core::hash::Hashed;
-use grin_core::core::id::ShortIdentifiable;
 use grin_core::core::verifier_cache::{LruVerifierCache, VerifierCache};
 use grin_core::core::Committed;
-use grin_core::core::{Block, BlockHeader, CompactBlock, KernelFeatures, OutputFeatures};
+use grin_core::core::{Block, BlockHeader, KernelFeatures, OutputFeatures};
 use grin_core::{global, ser};
 use keychain::{BlindingFactor, ExtKeychain, Keychain};
 use util::{secp, secp_static};
@@ -282,33 +280,6 @@ fn block_single_tx_serialized_size() {
 }
 
 #[test]
-fn empty_compact_block_serialized_size() {
-	let keychain = ExtKeychain::from_random_seed().unwrap();
-	let prev = BlockHeader::default();
-	let key_id = keychain.derive_key_id(1).unwrap();
-	let b = new_block(vec![], &keychain, &prev, &key_id);
-	let cb: CompactBlock = b.into();
-	let mut vec = Vec::new();
-	ser::serialize(&mut vec, &cb).expect("serialization failed");
-	let target_len = 1_260;
-	assert_eq!(vec.len(), target_len);
-}
-
-#[test]
-fn compact_block_single_tx_serialized_size() {
-	let keychain = ExtKeychain::from_random_seed().unwrap();
-	let tx1 = tx1i2o();
-	let prev = BlockHeader::default();
-	let key_id = keychain.derive_key_id(1).unwrap();
-	let b = new_block(vec![&tx1], &keychain, &prev, &key_id);
-	let cb: CompactBlock = b.into();
-	let mut vec = Vec::new();
-	ser::serialize(&mut vec, &cb).expect("serialization failed");
-	let target_len = 1_266;
-	assert_eq!(vec.len(), target_len);
-}
-
-#[test]
 fn block_10_tx_serialized_size() {
 	let keychain = ExtKeychain::from_random_seed().unwrap();
 	global::set_mining_mode(global::ChainTypes::Mainnet);
@@ -325,116 +296,6 @@ fn block_10_tx_serialized_size() {
 	ser::serialize(&mut vec, &b).expect("serialization failed");
 	let target_len = 17_072;
 	assert_eq!(vec.len(), target_len,);
-}
-
-#[test]
-fn compact_block_10_tx_serialized_size() {
-	let keychain = ExtKeychain::from_random_seed().unwrap();
-
-	let mut txs = vec![];
-	for _ in 0..10 {
-		let tx = tx1i2o();
-		txs.push(tx);
-	}
-	let prev = BlockHeader::default();
-	let key_id = keychain.derive_key_id(1).unwrap();
-	let b = new_block(txs.iter().collect(), &keychain, &prev, &key_id);
-	let cb: CompactBlock = b.into();
-	let mut vec = Vec::new();
-	ser::serialize(&mut vec, &cb).expect("serialization failed");
-	let target_len = 1_320;
-	assert_eq!(vec.len(), target_len,);
-}
-
-#[test]
-fn compact_block_hash_with_nonce() {
-	let keychain = ExtKeychain::from_random_seed().unwrap();
-	let tx = tx1i2o();
-	let prev = BlockHeader::default();
-	let key_id = keychain.derive_key_id(1).unwrap();
-	let b = new_block(vec![&tx], &keychain, &prev, &key_id);
-	let cb1: CompactBlock = b.clone().into();
-	let cb2: CompactBlock = b.clone().into();
-
-	// random nonce will not affect the hash of the compact block itself
-	// hash is based on header POW only
-	assert!(cb1.nonce != cb2.nonce);
-	assert_eq!(b.hash(), cb1.hash());
-	assert_eq!(cb1.hash(), cb2.hash());
-
-	assert!(cb1.kern_ids()[0] != cb2.kern_ids()[0]);
-
-	// check we can identify the specified kernel from the short_id
-	// correctly in both of the compact_blocks
-	assert_eq!(
-		cb1.kern_ids()[0],
-		tx.kernels()[0].short_id(&cb1.hash(), cb1.nonce)
-	);
-	assert_eq!(
-		cb2.kern_ids()[0],
-		tx.kernels()[0].short_id(&cb2.hash(), cb2.nonce)
-	);
-}
-
-#[test]
-fn convert_block_to_compact_block() {
-	let keychain = ExtKeychain::from_random_seed().unwrap();
-	let tx1 = tx1i2o();
-	let prev = BlockHeader::default();
-	let key_id = keychain.derive_key_id(1).unwrap();
-	let b = new_block(vec![&tx1], &keychain, &prev, &key_id);
-	let cb: CompactBlock = b.clone().into();
-
-	assert_eq!(cb.out_full().len(), 1);
-	assert_eq!(cb.kern_full().len(), 1);
-	assert_eq!(cb.kern_ids().len(), 1);
-
-	assert_eq!(
-		cb.kern_ids()[0],
-		b.kernels()
-			.iter()
-			.find(|x| !x.features.contains(KernelFeatures::COINBASE_KERNEL))
-			.unwrap()
-			.short_id(&cb.hash(), cb.nonce)
-	);
-}
-
-#[test]
-fn hydrate_empty_compact_block() {
-	let keychain = ExtKeychain::from_random_seed().unwrap();
-	let prev = BlockHeader::default();
-	let key_id = keychain.derive_key_id(1).unwrap();
-	let b = new_block(vec![], &keychain, &prev, &key_id);
-	let cb: CompactBlock = b.clone().into();
-	let hb = Block::hydrate_from(cb, vec![]).unwrap();
-	assert_eq!(hb.header, b.header);
-	assert_eq!(hb.outputs(), b.outputs());
-	assert_eq!(hb.kernels(), b.kernels());
-}
-
-#[test]
-fn serialize_deserialize_compact_block() {
-	let keychain = ExtKeychain::from_random_seed().unwrap();
-	let tx1 = tx1i2o();
-	let prev = BlockHeader::default();
-	let key_id = keychain.derive_key_id(1).unwrap();
-	let b = new_block(vec![&tx1], &keychain, &prev, &key_id);
-
-	let mut cb1: CompactBlock = b.into();
-
-	let mut vec = Vec::new();
-	ser::serialize(&mut vec, &cb1).expect("serialization failed");
-
-	// After header serialization, timestamp will lose 'nanos' info, that's the designed behavior.
-	// To suppress 'nanos' difference caused assertion fail, we force b.header also lose 'nanos'.
-	let origin_ts = cb1.header.timestamp;
-	cb1.header.timestamp =
-		origin_ts - Duration::nanoseconds(origin_ts.timestamp_subsec_nanos() as i64);
-
-	let cb2: CompactBlock = ser::deserialize(&mut &vec[..]).unwrap();
-
-	assert_eq!(cb1.header, cb2.header);
-	assert_eq!(cb1.kern_ids(), cb2.kern_ids());
 }
 
 #[test]
